requirements.txt
Pillow == 8.0.1
PyQt5_sip == 12.8.1
